requirements.txt
chromadb
fastapi
langchain
langchain-community
langchain-groq
langgraph
playwright
pypdf
python-multipart
python-dotenv
uvicorn
